moved constants to the connectors
