+++ v2/datastream-to-spanner/src/main/java/com/google/cloud/teleport/v2/templates/source/postgresql/PostgresChangeEventSequence.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.templates.datastream.source.postgresql;
+package com.google.cloud.teleport.v2.templates.source.postgresql;
 
 import com.google.cloud.spanner.Key;
 import com.google.cloud.spanner.ResultSet;
