+++ v2/datastream-to-spanner/src/main/java/com/google/cloud/teleport/v2/templates/source/postgresql/package-info.java
@@ -15,4 +15,4 @@
  */
 
 /** PostgreSQL source connector implementation. */
-package com.google.cloud.teleport.v2.templates.datastream.source.postgresql;
+package com.google.cloud.teleport.v2.templates.source.postgresql;
