+++ v2/datastream-to-spanner/src/test/java/com/google/cloud/teleport/v2/templates/SpannerStreamingWriteIntegrationTest.java
@@ -30,7 +30,9 @@ import com.google.cloud.teleport.v2.spanner.SpannerServerResource;
 import com.google.cloud.teleport.v2.spanner.ddl.Ddl;
 import com.google.cloud.teleport.v2.spanner.ddl.InformationSchemaScanner;
 import com.google.cloud.teleport.v2.spanner.migrations.schema.Schema;
+import com.google.cloud.teleport.v2.spanner.source.SourceConstants;
 import com.google.cloud.teleport.v2.templates.datastream.DatastreamConstants;
+import com.google.cloud.teleport.v2.templates.source.oracle.OracleDsToSpSourceConnector;
 import com.google.cloud.teleport.v2.templates.spanner.ProcessInformationSchema;
 import com.google.cloud.teleport.v2.values.FailsafeElement;
 import java.util.Arrays;
@@ -177,11 +179,11 @@ public class SpannerStreamingWriteIntegrationTest {
 
   private JSONObject getChangeEvent(String tableName, String changeType, String scn) {
     JSONObject json = new JSONObject();
-    json.put(DatastreamConstants.EVENT_SOURCE_TYPE_KEY, DatastreamConstants.ORACLE_SOURCE_TYPE);
+    json.put(DatastreamConstants.EVENT_SOURCE_TYPE_KEY, SourceConstants.ORACLE_SOURCE_TYPE);
     json.put(DatastreamConstants.EVENT_CHANGE_TYPE_KEY, changeType);
     json.put(DatastreamConstants.EVENT_TABLE_NAME_KEY, tableName);
-    json.put(DatastreamConstants.ORACLE_TIMESTAMP_KEY, 1);
-    json.put(DatastreamConstants.ORACLE_SCN_KEY, scn);
+    json.put(OracleDsToSpSourceConnector.ORACLE_TIMESTAMP_KEY, 1);
+    json.put(OracleDsToSpSourceConnector.ORACLE_SCN_KEY, scn);
     return json;
   }
