+++ v2/datastream-to-spanner/src/test/java/com/google/cloud/teleport/v2/templates/source/mysql/MySqlChangeEventContextTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.templates.datastream.source.mysql;
+package com.google.cloud.teleport.v2.templates.source.mysql;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
