+++ v2/datastream-to-spanner/src/test/java/com/google/cloud/teleport/v2/templates/source/postgresql/PostgresChangeEventSequenceTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.templates.datastream.source.postgresql;
+package com.google.cloud.teleport.v2.templates.source.postgresql;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
