+++ v2/spanner-common/src/main/java/com/google/cloud/teleport/v2/spanner/source/SourceConstants.java
@@ -1,5 +1,45 @@
+/*
+ * Copyright (C) 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package com.google.cloud.teleport.v2.spanner.source;
 
+/*
+ * Copyright (C) 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+/** Constants related to supported sources for spanner migrations. */
 public class SourceConstants {
 
+  /* The source type value for MySql databases */
+  public static final String MYSQL_SOURCE_TYPE = "mysql";
+
+  /* The value for Oracle databases in the source type key */
+  public static final String ORACLE_SOURCE_TYPE = "oracle";
+
+  /* The value for Postgres databases in the source type key */
+  public static final String POSTGRES_SOURCE_TYPE = "postgresql";
 }
